First cut at implementing processes within muSE. It seems to work beautifully! The mechanism is very simple - a process is given a certain attention span and when that attention span expires, the run loop switches to the next process.
1. Facility to spawn new processes.
2. Facility to mark atomic blocks during which
   process switch shouldn't occur. 
3. All symbols' values are now process local. (MAJOR)
4. Each process has an Erlang style mailbox using which
   it can send and receive messages to other processes.
   Functionality not fleshed out entirely.

The purpose of this commit is to put in a version
in which the basics work. For example, you can spawn
a few processes and watch them run in the REPL. Of course,
the REPL itself will be given attention and will end up
blocking the other processes. When that happens, evaluate
some expression that will exhaust the REPL's attention and
the other processes will get a chance to run.