Added process primitives "post" and "process?".
"post" lets you post arbitrary messages to processes.
In particular this is useful to postpone the processing
of a message if it cannot be handled upon reception.

"process?" helps you check whether the given object is 
a pid. If so, it evaluates to the pid itself as the
result. If not, it evaluates to ().