Modified the semantics of muse_assoc_iter().
It used to return NULL when an item is not found. Instead, it now returns a pointer to the tail cell of the list being searched. This way, if we decide to add another element to the assoc list, we can immediately add it to the end of the list.

The hashtable code was affected, but hashtable modification should be faster as a result of the changes.