FEATURE: (experimental)
Introduced "it" (which is now a reserved symbol). In conditional
expressions such as if, when, unles, cond and case, the value of
the condition on which the branch occurs will be available in the
body of the conditional as "it". For example, 

(if (find 3 '(1 2 3 4 5))
   (print it)
   (print "not found"))

will print 

(3 4 5)

This is marked "experimental" since "it" may not be the right word
for it. Maybe "the-result" is better, but "it" seems the most concise.