FEATURE: When "import" is used at the top level, it first "require"s all the modules and then imports them. This simplifies module usage so that instead of writing   (require XML)   (import XML) you can simply write   (import XML)
"import" won't do a require when it is used in a local scope. This
allows the use of modules in function arguments.

BUGFIX: Previously, when import processes the first argument, the
resulting bindings can influence the evaluation of subsequent arguments.
This is usually undesirable. The new implementation evaluates all
arguments first before processing them.